pyproject.toml
[project]
name = "tide"
version = "0.1.0"
description = "A Zenoh-based robotics framework with opinionated namespacing"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "eclipse-zenoh>=1.3.4",
    "hypothesis>=6.131.15",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
]

[project.scripts]
tide = "main:main"
